The code changes involve updates to both the database schema and the user interface of a .NET Core application. The database schema changes are implemented through several migration files, which alter various properties in the `TreatmentRecords`, `Measurements`, and `BreedingRecords` tables to be nullable. This means that these fields are no longer required in the database.
The user interface changes involve updates to several Razor files, adding new features, fixing bugs, and improving the user experience. These changes include adding support for a new type of duty, `Treatment`, in the duty editor component, adding a new navigation link for `Treatments` in the main navigation menu, fixing a bug in the label for the male selection dropdown in the `BreedingRecordEditor`, and adding a new column for `Location` in the livestock grid.

The code changes also involve refactoring and enhancing the functionality of two components: `TreatmentEditor.razor` and `TreatmentRecordEditor.razor`. These components now include additional fields and methods for handling units and treatments.

The changes are as follows:

1. [MicroAgManagementDbContextModelSnapshot.cs](#)
2. [DutyConstants.cs, MeasurementConstants.cs, TreatmentConstants.cs](#)
3. [BreedingRecord.cs, Measurement.cs, TreatmentRecord.cs](#)
4. [Chore.cs, ChoreModel.cs](#)
5. [MeasurementModel.cs, TreatmentRecordModel.cs](#)
6. [FrontEndDbContextModelSnapshot.cs](#)
7. [DutyEditor.razor, MainNavMenu.razor, BreedingRecordEditor.razor, LivestockGrid.razor, MeasureEditor.razor, MeasurementGrid.razor, RegistrationEditor.razor, ScheduledDutyEditor.razor](#)
8. [TreatmentEditor.razor, TreatmentRecordEditor.razor](#)
9. [TreatmentRecordGrid.razor, BreedingRecords.razor, Land.razor, Livestock.razor](#)
10. [20240309174112_Treatments.cs, 20240309174623_Treatments2.cs, 20240309175332_Treatments3.cs, 20240309182844_Measurement.cs, 20240309185035_BreedingRecord.cs](#)
11. [20240309174235_Treatments.cs, 20240309174556_Treatments2.cs, 20240309175238_Treatments3.cs, 20240309182811_Measurement.cs, 20240309184907_BreedingRecord.cs](#)
12. [TreatmentRecords.razor](#)