The most significant changes in the code include the modification of the project type guid in the "MicroAgManager.sln" solution file, the removal of several "using" directives from multiple files, and the renaming of the "SourceIds" method to "ScheduledDutySourceIds" in the "DutyLogic.cs" file. Additionally, the logic of the application has been altered with the addition and removal of several lines of code in the "DutyLogic.cs" file.
Other notable changes include the removal of the "Domain.Models" namespace from the "TestMicroAgManagementDbContext.cs" file, the modification of table names in the "FrontEndDbContextModelSnapshot.cs" file, and the complete removal of the "Login.razor" and "Register.razor" files. The "Logout.razor" file has been wrapped in a form tag, and the layout of the main page has been changed with updates to the "MainLayout.razor.css", "MainLayout.razor.min.css", and "MainLayout.razor.scss" files.

The "ScheduledDutyEditor.razor" file has been updated to use the renamed "ScheduledDutySourceIds" method, and a new file "IdentityForm.razor.scss" has been added to the "compilerconfig.json" file. The "#app" and "form" selectors in the "app.css" file have been updated to use grid display and center items, and change the layout of forms, respectively.

Finally, the "i.fa-glass-water-droplet" and "form .updateCancelButtons" selectors in the "app.css" file have been updated to change the layout of specific elements. The changes in the "app.scss" file include modifications to the height property for "html, body" and "#app", and updates to the "#app" and "form" selectors to use CSS Grid for layout.

New unit tests have been added for the "LivestockAnimalLogic" and "LivestockLogic" classes in the "LivestockAnimalLogicTests.cs" and "LivestockLogicTests.cs" files, respectively. A new Blazor component for user authentication has been added in the "IdentityForm.razor" file, and new CSS styles have been added for the "IdentityForm" component in the "IdentityForm.razor.css" and "IdentityForm.razor.min.css" files. The SCSS source code for the "IdentityForm" component's styles has been added in the "IdentityForm.razor.scss" file.